services/importer/requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
msoffcrypto-tool>=5.4.2
xlrd>=2.0.1
